Password change validations moved to helper.js
